requirements.txt
beautifulsoup4==4.9.0
boto3==1.13.
dateparser==0.7.4
python-dotenv==0.13.0
requests==2.23.0
